requirements.txt
dash==2.10.2
dash-cytoscape==0.3.0
dash-bootstrap-components==1.5.0
dash-mantine-components==0.12.0
matplotlib==3.7.1
numpy==1.24.2
pandas==2.0.1
plotly==5.17.0
plotly-express==0.4.1
python-dotenv==1.0.0
dash-iconify==0.1.2